Main Method
Console app created using Visual Studio in C# demonstrates class creation and method overload. Main method takes an integer and returns result as an integer. Second method takes in decimals, performs a math operation, and returns as an integer. Third method takes a string and converts to integer if possible performing a math operation and returns answer as an integer. Main() method of console app instantiates an object of class created and displays results.